Switch tutorial to Literate.jl.
