+++ .gitlab-ci.yml
@@ -41,31 +41,10 @@ documentation:
     - staging
     - trying
 
-tutorial:
-  stage: post
-  image: "juliagpu/julia:v1.1-${CI_IMAGE_TAG}"
-  variables:
-    JULIA_NUM_THREADS: "4"
-  dependencies:
-    - test:v1.1
-  script:
-    - julia --project=tutorials/ -e 'using Pkg;
-                                     Pkg.develop(PackageSpec(path=pwd()));
-                                     Pkg.instantiate()'
-    - julia --project=tutorials/ tutorials/make.jl
-  artifacts:
-    paths:
-    - tutorials/build
-  only:
-    - master
-    - staging
-    - trying
-
 pages:
   stage: deploy
   script:
     - mv docs/build public
-    - mv tutorials/build public/tutorials
   artifacts:
     paths:
     - public
